Move device-related files into source/device, logger-related into source/logger
