+++ source/device/switch.cpp
